Separate the environment by \0 (Fixes #36)
Output the environment from the shell separated by \0 instead of \n. This prevents confusion when the value of an environment variable contains newlines. (The value of an environment variable can't contain \0.)

The sh script is pretty ugly, but it should work everywhere. Unfortunately ``printenv --null`` is not available on many platforms.